pyproject.toml
[project]
name = "visual-buffet"
version = "0.1.10"
description = "Visual Buffet - Compare visual tagging results from local ML tools"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Bryant" }
]
keywords = ["image", "machine-learning", "tagging", "cli", "comparison"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pillow>=10.0",
    "pillow-heif>=0.16",
    "psutil>=5.9",
    "tomli-w>=1.0",
    "tomli>=2.0;python_version<'3.11'",
    "rawpy>=0.19",
    "packaging>=21.0",
    "paddlepaddle>=3.2.2",
    "paddleocr>=3.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.100",
    "jsonschema>=4.20",
    "ruff>=0.4",
    "httpx>=0.27",
]
gui = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-multipart>=0.0.9",
]
ml = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
]
ram_plus = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "einops>=0.7.0",
    "scipy>=1.10.0",
    "fairscale>=0.4.0",
    "ram @ git+https://github.com/xinyu1205/recognize-anything.git",
]
florence_2 = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.40.0,<4.50.0",
    "einops>=0.7.0",
]
siglip = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.47.0,<4.50.0",
    "accelerate>=0.25.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
]
siglip_quantized = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.47.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
]
yolo = [
    "torch>=2.0",
    "torchvision>=0.15",
    "ultralytics>=8.0",
]
easyocr = [
    "torch>=2.0",
    "torchvision>=0.15",
    "easyocr>=1.7.0",
]
# surya_ocr requires transformers>=4.56 which conflicts with RAM++ and other models
# Install surya-ocr manually if needed: uv pip install surya-ocr
# surya_ocr = [
#     "torch>=2.0",
#     "torchvision>=0.15",
#     "surya-ocr>=0.17.0",
# ]
doctr = [
    "torch>=2.0",
    "torchvision>=0.15",
    "python-doctr>=0.11.0",
]
qwen3_vl = [
    "ollama>=0.4.0",
    "httpx>=0.25.0",
]
all_models = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "einops>=0.7.0",
    "scipy>=1.10.0",
    "fairscale>=0.4.0",
    "transformers>=4.47.0,<4.50.0",
    "accelerate>=0.25.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
    "ultralytics>=8.0",
    "easyocr>=1.7.0",
    "python-doctr>=0.11.0",
    "ram @ git+https://github.com/xinyu1205/recognize-anything.git",
]

[project.scripts]
visual-buffet = "visual_buffet.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Only resolve for the current platform (avoids cross-platform conflicts)
environments = ["sys_platform == 'linux'"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/visual_buffet"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
    "integration: marks integration tests",
    "regression: marks regression tests",
    "benchmark: marks benchmark tests",
]
addopts = "-v --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning:torch.*",
    "ignore::UserWarning:PIL.*",
]
